requirements.txt
numpy==1.21.2
Pillow==9.2.0
